requirements.txt
fastapi
uvicorn 
motor 
pydantic[dotenv]
numpy
transformers
sentence_transformers
pandas
requests
bs4
python-dotenv
sqlalchemy 
alembic 
openai==0.28
langchain-openai